newer? versions of octave, python, oct2py force an int type and pskmod is not happy

+++ simulink/o_rrc_mod.m
@@ -4,7 +4,7 @@ function [ txSig ] = o_rrc_mod(data, rrcFilter, rolloff, span, sps, M )
 
 % rrcFilter = rcosdesign(rolloff, span, sps);
 
-modData = pskmod(data, M, pi/4);
+modData = pskmod(data, double(M), pi/4);
 
 txSig = upfirdn(modData, rrcFilter, sps, 1);
